cargo-insta: use cargo integration tests
