+++ cargo-insta/Cargo.toml
@@ -23,3 +23,6 @@ syn = { version = "1.0.50", features = ["full", "visit", "extra-traits"] }
 ignore = "0.4.17"
 uuid = { version = "1.0.0", features = ["v4"] }
 tempfile = "3.5.0"
+
+[dev-dependencies]
+walkdir = "2.3.1"
