--- cargo-insta/integration-tests/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "integration-tests"
-version = "0.1.0"
-authors = ["Armin Ronacher <armin.ronacher@active-4.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dircpy = "0.3.4"
-insta = { path = "../..", features = ["json", "yaml", "redactions", "glob"] }
-walkdir = "2.3.1"
-serde = { version = "1.0.117", features = ["derive"] }
