+++ cargo-insta/tests/test-input/test_basic_utf8_inline.rs
