+++ cargo-insta/tests/test-input/test_json_inline.rs
