+++ cargo-insta/tests/test-input/test_yaml_inline.rs
